Naming things: Rename package to `tikray`
`loko` could not be uploaded to PyPI, because its name was too similar
to other projects on PyPI.

> The project name is too similar to an existing project and may be
> confusable.
> -- https://pypi.org/help/#project-name
