+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Unreleased
+- Renamed package to `tikray`
 
 ## 2025/02/05 v0.0.23
 - Renamed package to `loko`
